requirements.txt
django==4.1.13
kafka-python==2.0.2
djangorestframework==3.14.0
pymongo==3.11.4
scikit-learn==1.3.2
newsapi-python
pytz
djongo==1.3.6
dnspython==2.6.1
cryptography==44.0.0